Added the interactive functionality to the main application.
