+++ index.js
@@ -9,13 +9,17 @@ var colors = require('colors');
 var OnboardingCli = require("./onboardingCli");
 var TranslatorCli = require("./translatorCli");
 var fs = require('fs');
+var inquirer = require('inquirer');
+var q = require('q');
 var helpers = require('./helpers');
 var translatorCli = new TranslatorCli();
+var MainController = require("./controllers/mainController");
 
 // set theme 
 colors.setTheme({
   silly: 'rainbow',
-  header: 'cyan'
+  header: 'cyan',
+  state: 'yellow'
 });
 
 program
@@ -30,12 +34,17 @@ program
     .option('-g --get [RAML property name]', 'Property name to GET for -t')
     .option('-s --set [RAML property name]', 'Property name to SET for -t')
     .option('-v --value [value]', 'Stringified JSON value to pass in')
+    .option('-m --menu', 'Launch interactive menu')
     .parse(process.argv);
 
 console.log('Open Translators to Things CLI:');
 console.log('');
 
-if (program.onboarding) {
+if(program.menu) {
+    prompt({ currentController: new MainController(), controllerStack: [] });
+}
+
+else if (program.onboarding) {
     console.log("------ Doing onboarding for %j".header, program.onboarding);
     
     var fileName = helpers.createOnboardingFileName(program.onboarding);
@@ -185,4 +194,47 @@ else {
 function make_red(txt) {
     //display the help text in red on the console
     return colors.red(txt);
+}
+
+function prompt(state) {
+    state.currentController.logState(state);
+    let operations = state.currentController.getOperations(state);
+
+    let question = {
+        type: 'rawlist',
+        name: 'choice',
+        message: 'What would you like to do?',
+        choices: operations.map(o => o.name)
+    };
+
+    inquirer.prompt([question]).then(function (answers) {
+        if (answers.choice !== 'Exit') {
+            let choice = operations.find(o => o.name === answers.choice);
+            if (choice !== undefined) {
+                process.stdout.write('\033c');
+                choice.operation(state).then(prompt)
+                    .catch(err => {
+                        helpers.logError(err);
+                        prompt(state);
+                    });
+            }
+            else {
+                helpers.logError('Unknown operation');
+                prompt(state);
+            }
+        }
+        else {
+            let question = {
+                type: 'confirm',
+                name: 'exit',
+                message: 'Are you sure you want to exit?'
+            };
+
+            inquirer.prompt([question]).then(function (answers) {
+                if(!answers.exit) {
+                    prompt(state);
+                }
+            });
+        }
+    });
 }
